LayoutTests/platform/glib/imported/w3c/web-platform-tests/cookie-store/change_eventhandler_for_no_name_multiple_values.https.window-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: The cookie name and value must not both be empty.

FAIL Verify behavior of multiple no-name cookies promise_test: Unhandled rejection with value: object "TypeError: The cookie name and value must not both be empty."
